add deep compare method for absd.
the source class can define == operator or is_eq method. if two objects
has same classes the operator or method will be used to compare.

gcc has a bug with dynamic_cast - it returns nullptr in all cases. so
the workaround has beed added: we can compare pointers to strings with
__PRETTY_FUNCTION__ to get type info (it contains calss name with all
template parameters).