add move ctor and factory usage to virtual variant
add tests and fix to the variant on move ctor and move operator= - the
pointer should to be setted correctly

also each item will be created with a factory and add tests what each
item uses copy of original factory. it allow us to tune factory and use
it in all variant elements.

hide pointer - the virtual_variant::solve method should to be used (the
pointer is a invariant).

add exec method instead of visit. the exec methods loose returning
value. also it allows use clang workaround without visit.

methods as visit and holds_alternative should to be called on holder
member.