bock_set setts result with content
the holder type now just call the execute method on each owned object.

the block_set just create a string from content and set the string as value of stack variable. it can work better if the execute method use the result as is, not as a string. it can be achieved in block_set with braces.